pyproject.toml
[tool.poetry]
name = "cdk-cloudwatch-custom-metrics"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.109.0"
python-dotenv = "^0.17.1"
black = "^21.6b0"
isort = "^5.8.0"
flake8 = "^3.9.2"
"aws-cdk.aws-cloudwatch" = "^1.109.0"
"aws-cdk.aws-lambda" = "^1.109.0"
"aws-cdk.aws-events" = "^1.109.0"
"aws-cdk.aws-events-targets" = "^1.109.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
